+++ test/evaluator.test.ts
@@ -1,8 +1,5 @@
-import * as fs from 'fs';
 import glob from 'glob';
-import * as path from 'path';
 import { evaluate, generateVarCombinations, isAllowedPrompt } from '../src/evaluator';
-import { renderPrompt, resolveVariables } from '../src/evaluatorHelpers';
 import type { ApiProvider, TestSuite, Prompt } from '../src/types';
 
 jest.mock('node-fetch', () => jest.fn());
@@ -1229,148 +1226,6 @@ describe('evaluator', () => {
       }),
       expect.anything(),
     );
-  });
-});
-
-describe('renderPrompt', () => {
-  it('should render a prompt with a single variable', async () => {
-    const prompt = toPrompt('Test prompt {{ var1 }}');
-    const renderedPrompt = await renderPrompt(prompt, { var1: 'value1' }, {});
-    expect(renderedPrompt).toBe('Test prompt value1');
-  });
-
-  it('should render a JSON prompt', async () => {
-    const prompt = toPrompt('[{"text": "Test prompt "}, {"text": "{{ var1 }}"}]');
-    const renderedPrompt = await renderPrompt(prompt, { var1: 'value1' }, {});
-    expect(renderedPrompt).toBe(
-      JSON.stringify(JSON.parse('[{"text":"Test prompt "},{"text":"value1"}]'), null, 2),
-    );
-  });
-
-  it('should render a JSON prompt and escape the var string', async () => {
-    const prompt = toPrompt('[{"text": "Test prompt "}, {"text": "{{ var1 }}"}]');
-    const renderedPrompt = await renderPrompt(prompt, { var1: 'He said "hello world!"' }, {});
-    expect(renderedPrompt).toBe(
-      JSON.stringify(
-        JSON.parse('[{"text":"Test prompt "},{"text":"He said \\"hello world!\\""}]'),
-        null,
-        2,
-      ),
-    );
-  });
-
-  it('should render a JSON prompt with nested JSON', async () => {
-    const prompt = toPrompt('[{"text": "Test prompt "}, {"text": "{{ var1 }}"}]');
-    const renderedPrompt = await renderPrompt(prompt, { var1: '{"nested": "value1"}' }, {});
-    expect(renderedPrompt).toBe(
-      JSON.stringify(
-        JSON.parse('[{"text":"Test prompt "},{"text":"{\\"nested\\": \\"value1\\"}"}]'),
-        null,
-        2,
-      ),
-    );
-  });
-
-  it('should load external yaml files in renderPrompt', async () => {
-    const prompt = toPrompt('Test prompt with {{ var1 }}');
-    const vars = { var1: 'file://test.txt' };
-    const evaluateOptions = {};
-
-    // Mock fs.readFileSync to simulate loading a YAML file
-    jest.spyOn(fs, 'readFileSync').mockReturnValueOnce('loaded from file');
-
-    const renderedPrompt = await renderPrompt(prompt, vars, evaluateOptions);
-
-    expect(fs.readFileSync).toHaveBeenCalledWith(expect.stringContaining('test.txt'), 'utf8');
-    expect(renderedPrompt).toBe('Test prompt with loaded from file');
-  });
-
-  it('should load external js files in renderPrompt and execute the exported function', async () => {
-    const prompt = toPrompt('Test prompt with {{ var1 }} {{ var2 }} {{ var3 }}');
-    const vars = {
-      var1: 'file:///path/to/testFunction.js',
-      var2: 'file:///path/to/testFunction.cjs',
-      var3: 'file:///path/to/testFunction.mjs',
-    };
-    const evaluateOptions = {};
-
-    jest.doMock(
-      path.resolve('/path/to/testFunction.js'),
-      () => (varName: any, prompt: any, vars: any) => ({ output: `Dynamic value for ${varName}` }),
-      { virtual: true },
-    );
-    jest.doMock(
-      path.resolve('/path/to/testFunction.cjs'),
-      () => (varName: any, prompt: any, vars: any) => ({ output: `and ${varName}` }),
-      { virtual: true },
-    );
-    jest.doMock(
-      path.resolve('/path/to/testFunction.mjs'),
-      () => (varName: any, prompt: any, vars: any) => ({ output: `and ${varName}` }),
-      { virtual: true },
-    );
-
-    const renderedPrompt = await renderPrompt(prompt, vars, evaluateOptions);
-    expect(renderedPrompt).toBe('Test prompt with Dynamic value for var1 and var2 and var3');
-  });
-
-  it('should load external json files in renderPrompt and parse the JSON content', async () => {
-    const prompt = toPrompt('Test prompt with {{ var1 }}');
-    const vars = { var1: 'file:///path/to/testData.json' };
-    const evaluateOptions = {};
-
-    jest.spyOn(fs, 'readFileSync').mockReturnValueOnce(JSON.stringify({ key: 'valueFromJson' }));
-
-    const renderedPrompt = await renderPrompt(prompt, vars, evaluateOptions);
-
-    expect(fs.readFileSync).toHaveBeenCalledWith(expect.stringContaining('testData.json'), 'utf8');
-    expect(renderedPrompt).toBe('Test prompt with {"key":"valueFromJson"}');
-  });
-
-  it('should load external yaml files in renderPrompt and parse the YAML content', async () => {
-    const prompt = toPrompt('Test prompt with {{ var1 }}');
-    const vars = { var1: 'file:///path/to/testData.yaml' };
-    const evaluateOptions = {};
-
-    jest.spyOn(fs, 'readFileSync').mockReturnValueOnce('key: valueFromYaml');
-
-    const renderedPrompt = await renderPrompt(prompt, vars, evaluateOptions);
-
-    expect(fs.readFileSync).toHaveBeenCalledWith(expect.stringContaining('testData.yaml'), 'utf8');
-    expect(renderedPrompt).toBe('Test prompt with {"key":"valueFromYaml"}');
-  });
-});
-
-describe('resolveVariables', () => {
-  it('should replace placeholders with corresponding variable values', () => {
-    const variables = { final: '{{ my_greeting }}, {{name}}!', my_greeting: 'Hello', name: 'John' };
-    const expected = { final: 'Hello, John!', my_greeting: 'Hello', name: 'John' };
-    expect(resolveVariables(variables)).toEqual(expected);
-  });
-
-  it('should handle nested variable substitutions', () => {
-    const variables = { first: '{{second}}', second: '{{third}}', third: 'value' };
-    const expected = { first: 'value', second: 'value', third: 'value' };
-    expect(resolveVariables(variables)).toEqual(expected);
-  });
-
-  it('should not modify variables without placeholders', () => {
-    const variables = { greeting: 'Hello, world!', name: 'John' };
-    const expected = { greeting: 'Hello, world!', name: 'John' };
-    expect(resolveVariables(variables)).toEqual(expected);
-  });
-
-  it('should not fail if a variable is not found', () => {
-    const variables = { greeting: 'Hello, {{name}}!' };
-    expect(resolveVariables(variables)).toEqual({ greeting: 'Hello, {{name}}!' });
-  });
-
-  it('should not fail for unresolved placeholders', () => {
-    const variables = { greeting: 'Hello, {{name}}!', name: '{{unknown}}' };
-    expect(resolveVariables(variables)).toEqual({
-      greeting: 'Hello, {{unknown}}!',
-      name: '{{unknown}}',
-    });
   });
 });
